requirements.txt
python-dotenv
requests
pywebview
pygobject